+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- [Add Output Destinations configuration](https://github.com/BetterThanTomorrow/calva/issues/1104)
+
 ## [2.0.422] - 2024-03-11
 
 - Bump deps.clj to `v1.11.2.1446`